refactor: 숫자 생성 로직 추상화
- NumberGenerator - RandomNumberGenerator
