+++ src/main/java/racingcar/domain/NumberGenerator.java
@@ -1,13 +1,5 @@
 package racingcar.domain;
 
-import camp.nextstep.edu.missionutils.Randoms;
-
-public class NumberGenerator {
-
-	private static final int MIN_NUMBER = 0;
-	private static final int MAX_NUMBER = 9;
-
-	public int generateNumber() {
-		return Randoms.pickNumberInRange(MIN_NUMBER, MAX_NUMBER);
-	}
+public interface NumberGenerator {
+	int generateNumber();
 }
